+++ Lunar/DataModels/BlockResponseModel.swift
@@ -1,5 +1,5 @@
 //
-//  BlockUserResponseModel.swift
+//  BlockResponseModel.swift
 //  Lunar
 //
 //  Created by Mani on 07/12/2023.
@@ -7,12 +7,14 @@
 
 import Foundation
 
-struct BlockUserResponseModel: Codable {
+struct BlockResponseModel: Codable {
   let person: PersonObject?
+  let community: CommunityObject?
   let blocked: Bool
 
   enum CodingKeys: String, CodingKey {
     case person = "person_view"
+    case community = "community_view"
     case blocked
   }
 }