+++ CHANGELOG.md
@@ -15,12 +15,14 @@ All notable changes to this project will be documented in this file.
 - Reporting posts
 - Blocking users via a comment
 - Reporting comments
+- Blocking communities
 - Warning to notify users about errors performing actions outside of home instance
 
 ### Misc
 
 - Removed `Shiny` package due to lag
 - Removed `Local Console` package
+- Cleaned up API fetchers code
 
 ## [2023.12.1]